33. 회원정보수정 - 완료

+++ src/main/java/com/cos/photogramstart/domain/user/User.java
@@ -8,6 +8,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 @Builder
@@ -19,7 +20,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // 번호 증가 전략이 데이터베이스를 따라간다.
-    private Long id;
+    private int id;
 
     @Column(length = 20, unique = true)
     private String username;
